third_party/WebKit/LayoutTests/fast/hidpi/static/mousewheel-scroll-amount.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<script src="../../../resources/js-test.js"></script>
<style type="text/css">
    body {
        height: 2000px;
        width: 2000px;
    }
</style>

<script>
    window.jsTestIsAsync = true;

    description("This test ensures that mouse wheel ticks scrolls the right number of pixels in high dpi mode.");

    function testDiagonalScroll() {
        // Reset
        document.scrollingElement.scrollTop = 0;
        document.scrollingElement.scrollLeft = 0;

        eventSender.mouseMoveTo(20, 20);
        // 1 tick should move 40px.
        eventSender.mouseScrollBy(-2, -2);

        shouldBecomeEqual("document.scrollingElement.scrollTop == 80 && " +
          "document.scrollingElement.scrollLeft == 80", "true", finishJSTest);
    }

    function runTest() {
        if (!window.eventSender || !window.internals) {
           debug("This test requires window.eventSender and window.internals.");
            finishJSTest();
            return;
        }

        // Turn on smooth scrolling.
        internals.settings.setScrollAnimatorEnabled(true);

        eventSender.mouseMoveTo(20, 20);
        // 1 tick should move 40px.
        eventSender.mouseScrollBy(-1, -1);

        // 40px per tick.
        shouldBecomeEqual("document.scrollingElement.scrollTop == 40 && " +
          "document.scrollingElement.scrollLeft == 40", "true", testDiagonalScroll);
    }
</script>

<body onload="runTest()"></body>
